backend_books_review/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
isort==5.13.2
Markdown==3.7
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
